Separate plan_pr into look_at_files and plan_pr
